Refatoração do modulo de criar projeto
