+++ src/main/java/com/br/estimativadeprojetodesoftware/view/projeto/CadastroProjetoView.java
@@ -1,0 +1,207 @@
+package com.br.estimativadeprojetodesoftware.view.projeto;
+
+import javax.swing.JButton;
+import javax.swing.JComboBox;
+import javax.swing.JList;
+import javax.swing.JScrollBar;
+import javax.swing.JScrollPane;
+import javax.swing.JTextField;
+
+/**
+ *
+ * @author tetzner
+ */
+public class CadastroProjetoView extends javax.swing.JInternalFrame {
+
+    /**
+     * Creates new form CadastroProjView
+     */
+    public CadastroProjetoView() {
+        initComponents();
+    }
+    
+    
+      public JButton getBtnAdicionarPerfil() {
+        return btnAdicionarPerfil;
+    }
+
+    public JButton getBtnCadastrarProjeto() {
+        return btnCadastrarProjeto;
+    }
+
+    public JButton getBtnCancelar() {
+        return btnCancelar;
+    }
+
+    public JComboBox<String> getCbmPerfis() {
+        return cbmPerfis;
+    }
+
+    public JList<String> getjList1() {
+        return jList1;
+    }
+    
+    public JScrollPane getSpListaPerfis() {
+        return spListaPerfis;
+    }
+
+    public JTextField getTxtNome() {
+        return txtNome;
+    }
+
+    public JTextField getTxtTipoProjeto() {
+        return txtTipoProjeto;
+    }  
+
+
+    /**
+     * This method is called from within the constructor to initialize the form.
+     * WARNING: Do NOT modify this code. The content of this method is always
+     * regenerated by the Form Editor.
+     */
+    @SuppressWarnings("unchecked")
+    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
+    private void initComponents() {
+
+        btnAdicionarPerfil = new javax.swing.JButton();
+        spListaPerfis = new javax.swing.JScrollPane();
+        jList1 = new javax.swing.JList<>();
+        txtNome = new javax.swing.JTextField();
+        txtTipoProjeto = new javax.swing.JTextField();
+        btnCadastrarProjeto = new javax.swing.JButton();
+        btnCancelar = new javax.swing.JButton();
+        lblCadastroProjeto = new javax.swing.JLabel();
+        lblNomeProjeto = new javax.swing.JLabel();
+        lblTipoProjeto = new javax.swing.JLabel();
+        lblPerfis = new javax.swing.JLabel();
+        cbmPerfis = new javax.swing.JComboBox<>();
+
+        setTitle("Cadastro de Projeto");
+
+        btnAdicionarPerfil.setText("Adicionar Perfil");
+
+        spListaPerfis.setViewportView(jList1);
+
+        txtNome.setMaximumSize(new java.awt.Dimension(64, 22));
+        txtNome.setName(""); // NOI18N
+
+        txtTipoProjeto.setMaximumSize(new java.awt.Dimension(64, 22));
+        txtTipoProjeto.setName(""); // NOI18N
+
+        btnCadastrarProjeto.setFont(new java.awt.Font("Segoe UI", 0, 14)); // NOI18N
+        btnCadastrarProjeto.setText("Cadastrar Projeto");
+
+        btnCancelar.setFont(new java.awt.Font("Segoe UI", 0, 14)); // NOI18N
+        btnCancelar.setText("Cancelar");
+        btnCancelar.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                btnCancelarActionPerformed(evt);
+            }
+        });
+
+        lblCadastroProjeto.setFont(new java.awt.Font("Segoe UI", 1, 24)); // NOI18N
+        lblCadastroProjeto.setText("Cadastro de Projeto");
+
+        lblNomeProjeto.setFont(new java.awt.Font("Segoe UI", 0, 14)); // NOI18N
+        lblNomeProjeto.setText("Nome:");
+
+        lblTipoProjeto.setFont(new java.awt.Font("Segoe UI", 0, 14)); // NOI18N
+        lblTipoProjeto.setText("Tipo:");
+
+        lblPerfis.setFont(new java.awt.Font("Segoe UI", 0, 14)); // NOI18N
+        lblPerfis.setText("Perfis:");
+
+        cbmPerfis.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                cbmPerfisActionPerformed(evt);
+            }
+        });
+
+        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
+        getContentPane().setLayout(layout);
+        layout.setHorizontalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(layout.createSequentialGroup()
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(layout.createSequentialGroup()
+                        .addGap(63, 63, 63)
+                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                            .addComponent(lblNomeProjeto)
+                            .addComponent(lblTipoProjeto)
+                            .addComponent(lblPerfis))
+                        .addGap(18, 18, 18)
+                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                            .addGroup(layout.createSequentialGroup()
+                                .addComponent(btnCadastrarProjeto)
+                                .addGap(18, 18, 18)
+                                .addComponent(btnCancelar, javax.swing.GroupLayout.PREFERRED_SIZE, 127, javax.swing.GroupLayout.PREFERRED_SIZE))
+                            .addGroup(layout.createSequentialGroup()
+                                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                                    .addComponent(spListaPerfis, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.PREFERRED_SIZE, 354, javax.swing.GroupLayout.PREFERRED_SIZE)
+                                    .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING, false)
+                                        .addComponent(cbmPerfis, 0, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                        .addComponent(txtTipoProjeto, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+                                        .addComponent(txtNome, javax.swing.GroupLayout.PREFERRED_SIZE, 354, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                                .addComponent(btnAdicionarPerfil))))
+                    .addGroup(layout.createSequentialGroup()
+                        .addGap(207, 207, 207)
+                        .addComponent(lblCadastroProjeto)))
+                .addContainerGap(44, Short.MAX_VALUE))
+        );
+        layout.setVerticalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(layout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(lblCadastroProjeto)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 18, Short.MAX_VALUE)
+                .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(layout.createSequentialGroup()
+                        .addComponent(lblNomeProjeto)
+                        .addGap(18, 18, 18)
+                        .addComponent(lblTipoProjeto)
+                        .addGap(18, 18, 18)
+                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                            .addComponent(lblPerfis)
+                            .addComponent(btnAdicionarPerfil)
+                            .addComponent(cbmPerfis, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                        .addGap(18, 18, 18)
+                        .addComponent(spListaPerfis, javax.swing.GroupLayout.PREFERRED_SIZE, 95, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addGap(18, 18, 18)
+                        .addGroup(layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                            .addComponent(btnCadastrarProjeto)
+                            .addComponent(btnCancelar)))
+                    .addGroup(layout.createSequentialGroup()
+                        .addComponent(txtNome, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                        .addGap(18, 18, 18)
+                        .addComponent(txtTipoProjeto, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                .addGap(19, 19, 19))
+        );
+
+        pack();
+    }// </editor-fold>//GEN-END:initComponents
+
+    private void btnCancelarActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnCancelarActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_btnCancelarActionPerformed
+
+    private void cbmPerfisActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_cbmPerfisActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_cbmPerfisActionPerformed
+
+
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JButton btnAdicionarPerfil;
+    private javax.swing.JButton btnCadastrarProjeto;
+    private javax.swing.JButton btnCancelar;
+    private javax.swing.JComboBox<String> cbmPerfis;
+    private javax.swing.JList<String> jList1;
+    private javax.swing.JLabel lblCadastroProjeto;
+    private javax.swing.JLabel lblNomeProjeto;
+    private javax.swing.JLabel lblPerfis;
+    private javax.swing.JLabel lblTipoProjeto;
+    private javax.swing.JScrollPane spListaPerfis;
+    private javax.swing.JTextField txtNome;
+    private javax.swing.JTextField txtTipoProjeto;
+    // End of variables declaration//GEN-END:variables
+}
